neccessary threads for stereo audio are all complete. need to verify if amplitude of signal need to be adjusted to achieve good stereo effect

+++ src/threads/mono_audio_extraction_thread.cpp
@@ -16,7 +16,7 @@ void* mono_audio_extraction_thread_diffeq(void* args){
 	m_audio_extract_args* params = (m_audio_extract_args*) args;
 	
 	BlockingQueue<double>* in = params->in;
-	BlockingQueue<float>* out = params->out;
+	BlockingQueue<double>* out = params->out;
 
 	double Fs = params->sample_rate;
     double audio_freq = 48000;
@@ -45,7 +45,7 @@ void* mono_audio_extraction_thread_diffeq(void* args){
     int counter = 0;    // every dec_rate sample, the sample is saved
     int index = 0;      // goes from 0 to chunk_size-1
 
-    float* extracted = new float[chunk_size];
+    double* extracted = new double[chunk_size];
     while(true){
         QueueElement<double>* popped = in->pop(10000, name);
         if(popped == nullptr){
@@ -87,7 +87,7 @@ void* mono_audio_extraction_thread_diffeq(void* args){
 
                 if(index == 0){
                     out->push(extracted);
-                    extracted = new float[chunk_size];
+                    extracted = new double[chunk_size];
                 }
             }
             counter = (counter+1) % dec_rate;
